Fixed unit test and convex hull implementation. Removed floating point rounding issues and adjusted min and max point.

+++ conv_hull/quick_hull/quick_hull.hpp
@@ -2,8 +2,8 @@
 #define QUICK_HULL
 
 typedef struct point{
-	double x;
-	double y;
+	long long int x;
+	long long int y;
 } Point;
 
 typedef Point Vector;
